--- .env.local
@@ -1,1 +1,0 @@
-NEXT_PUBLIC_BASE_URL =http://localhost:5003/api/
