+++ package.json
@@ -22,11 +22,12 @@
     "antd": "^5.1.2",
     "apexcharts": "^3.36.3",
     "axios": "^1.2.2",
+    "chart.js": "^4.1.2",
     "eslint": "8.30.0",
     "eslint-config-next": "13.1.1",
     "next": "13.1.1",
     "react": "18.2.0",
-    "react-apexcharts": "^1.4.0",
+    "react-chartjs-2": "^5.2.0",
     "react-dom": "18.2.0",
     "react-hook-form": "^7.41.5",
     "react-icons": "^4.7.1",
