+++ pages/chart.tsx
@@ -1,38 +1,73 @@
-import dynamic from "next/dynamic";
-import React from "react";
-import ReactApexChart, { Props } from "react-apexcharts";
+import { useState } from "react";
 
-const TasksChart: React.FC<Props> = () => {
-  const options = {
-    chart: {
-      height: 350,
-      zoom: {
-        enabled: true,
+import { Line } from "react-chartjs-2";
+import {
+  Chart as ChartJS,
+  Title,
+  Tooltip,
+  LineElement,
+  Legend,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  Filler,
+} from "chart.js";
+ChartJS.register(
+  Title,
+  Tooltip,
+  LineElement,
+  Legend,
+  CategoryScale,
+  LinearScale,
+  PointElement,
+  Filler
+);
+
+export default function TaskChart() {
+  const [data, setData] = useState({
+    labels: [
+      "Jan",
+      "Feb",
+      "March",
+      "April",
+      "May",
+      "June",
+      "July",
+      "August",
+      "September",
+      "Oct",
+      "Nov",
+      "Dec",
+    ],
+    datasets: [
+      {
+        label: "First Dataset",
+        data: [40, 50, 55, 60, 51, 82, 70, 59, 61, 73, 50, 30],
+        borderColor: "purple",
+        tension: 0.4,
+        fill: true,
+        pointStyle: "rect",
+        pointBorderColor: "blue",
+        pointBackgroundColor: "#fff",
+        showLine: true,
       },
-    },
-  };
-
-  const series = [
-    {
-      name: "All Tasks",
-      data: [31, 40, 28, 51, 42, 109, 100],
-    },
-    {
-      name: "My Tasks",
-      data: [11, 32, 45, 32, 34, 52, 41],
-    },
-  ];
+      {
+        label: "First Dataset",
+        data: [5, 5, 10, 10, 12, 14, 16, 18, 20, 30, 35, 40],
+        borderColor: "pink",
+        tension: 0.4,
+        fill: true,
+        pointStyle: "rect",
+        pointBorderColor: "blue",
+        pointBackgroundColor: "#fff",
+        showLine: true,
+      },
+    ],
+  });
 
   return (
     <>
-      <ReactApexChart
-        type="line"
-        options={options}
-        series={series}
-        height={305}
-      />
+      <Line data={data}></Line>
     </>
   );
-};
-
-export default TasksChart;
+}
